utils.c:humanize_bytes(): Avoid buffer overrun
Using static bufferpools as snprintf() targets makes a lot of sense,
but one must take care not to roll out of the buffer pool.

This patch adds the necessary bitmasking to make sure we stay inside
the safe zone and avoid any violations of the segmentation as we
do the buffer rotation.

Signed-off-by: Andreas Ericsson <ae@akapi.se>